Fix: moved user verified to a new resource, it will be common for discord and sms flows.
